fix(directedworkapi): fix 6 code review issues, extract directedworkapp
- Fix cycle count query: replace CountedBy filter with Status=pending
  (CountedBy is set after counting, not before — entire domain was dead code)
- Fix transfer query: split into approved (all workers) + in_transit
  (ClaimedByID filter excluded approved transfers with nil ClaimedByID)
- Fix DetailPath: use pick task ID instead of parent order ID
- Fix test: use transferorderbus.Status* constants instead of raw strings
- Fix test: add StatusRejected test case per documented contract
- Refactor: extract directedworkapp to app layer (CLAUDE.md: keep layers pure)
